202 merge to main

+++ src/mefit/Tools/Types/Structs.cs
@@ -12,7 +12,7 @@ namespace Mac_EFI_Toolkit.Tools
         internal string FileName { get; set; }
         internal string CreationTime { get; set; }
         internal string LastWriteTime { get; set; }
-        internal int Length { get; set; }
+        internal long Length { get; set; }
         internal uint CRC32 { get; set; }
     }
 }
